requirements.txt
Jinja2==3.1.4
MarkupSafe==2.1.5
netaddr==1.3.0
pyeapi==1.0.2
PyYAML==6.0.1
certifi==2024.6.2
charset-normalizer==3.3.2
cvprac==1.4.0
idna==3.7
Jinja2==3.1.4
MarkupSafe==2.1.5
netaddr==1.3.0
packaging==24.1
pyeapi==1.0.2
PySocks==1.7.1
PyYAML==6.0.1
requests==2.32.3
urllib3==2.2.2
